single_tech_samples/dataexplorer/src/python/requirements.txt
adal==1.2.4
azure-storage-blob==12.8.1
azure-storage-queue==12.1.6
azure-common==1.1.25
azure-kusto-data==2.1.1
azure-mgmt-kusto==0.9.0
azure-kusto-ingest==2.1.1
numpy==1.18.3